backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anyio==4.4.0
pyserial==3.5

opencv-python-headless==4.10.0.84
